Moved logging to separate file
To be reused by different modules.
